backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "surveillance"
version = "0.1.0"
description = "Student Surveillance System"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pydantic = "^2.5"
pydantic-settings = "^2.1"
numpy = "^1.24"
opencv-python = "^4.8"
pillow = "^10.1"
insightface = "^0.7.3"
onnxruntime = "^1.23"
faiss-cpu = "^1.7.4"
fastapi = "^0.104"
uvicorn = {extras = ["standard"], version = "^0.24"}
python-multipart = "^0.0.6"
# Enhanced features
pandas = "^2.1.0"
plotly = "^5.17.0"
# Optional enterprise features
sqlalchemy = {version = "^2.0", optional = true}
asyncpg = {version = "^0.29", optional = true}
alembic = {version = "^1.13", optional = true}
redis = {version = "^5.0", optional = true}
prometheus-client = {version = "^0.18", optional = true}
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
pytest-cov = "^4.1"
mypy = "^1.7"
types-redis = "^4.6"
ruff = "^0.1"
pre-commit = "^3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
